requirements.txt
# requirements.txt
apache-airflow
dotenv
pandas
psycopg2-binary
requests
